3. handle-db/project/requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
click==7.1.2
flake8==3.8.3
iniconfig==1.0.0
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.0.1
regex==2020.7.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1